refactor: extract shared widget logic into HasPeriodComparison trait
Deduplicate period calculation, change formatting, bucket expression,
sparkline filling, and active creator queries across ActivationRateWidget,
PlatformGrowthStatsWidget, and UserRetentionChartWidget.
