+++ app-modules/SystemAdmin/src/Filament/Widgets/PlatformGrowthStatsWidget.php
@@ -18,9 +18,11 @@ use Filament\Widgets\StatsOverviewWidget;
 use Filament\Widgets\StatsOverviewWidget\Stat;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Facades\DB;
+use Relaticle\SystemAdmin\Filament\Widgets\Concerns\HasPeriodComparison;
 
 final class PlatformGrowthStatsWidget extends StatsOverviewWidget
 {
+    use HasPeriodComparison;
     use InteractsWithPageFilters;
 
     protected static ?int $sort = 1;
@@ -44,20 +46,6 @@ final class PlatformGrowthStatsWidget extends StatsOverviewWidget
         ];
     }
 
-    /**
-     * @return array{0: CarbonImmutable, 1: CarbonImmutable, 2: CarbonImmutable, 3: CarbonImmutable}
-     */
-    private function getPeriodDates(): array
-    {
-        $days = (int) ($this->pageFilters['period'] ?? 30);
-        $currentEnd = CarbonImmutable::now();
-        $currentStart = $currentEnd->subDays($days);
-        $previousEnd = $currentStart->subSecond();
-        $previousStart = $currentStart->subDays($days);
-
-        return [$currentStart, $currentEnd, $previousStart, $previousEnd];
-    }
-
     private function buildUsersStat(
         CarbonImmutable $currentStart,
         CarbonImmutable $currentEnd,
@@ -73,7 +61,7 @@ final class PlatformGrowthStatsWidget extends StatsOverviewWidget
             ->description("{$newCurrent} new this period".$this->formatChange($change))
             ->descriptionIcon($change >= 0 ? 'heroicon-o-arrow-trending-up' : 'heroicon-o-arrow-trending-down')
             ->color($change >= 0 ? 'success' : 'danger')
-            ->chart($this->buildSparkline(User::class, $currentStart, $currentEnd));
+            ->chart($this->buildModelSparkline(User::class, $currentStart, $currentEnd));
     }
 
     private function buildTeamsStat(
@@ -93,7 +81,7 @@ final class PlatformGrowthStatsWidget extends StatsOverviewWidget
             ->description("{$newCurrent} new this period".$this->formatChange($change))
             ->descriptionIcon($change >= 0 ? 'heroicon-o-arrow-trending-up' : 'heroicon-o-arrow-trending-down')
             ->color($change >= 0 ? 'success' : 'danger')
-            ->chart($this->buildSparkline(Team::class, $currentStart, $currentEnd, $nonPersonalScope));
+            ->chart($this->buildModelSparkline(Team::class, $currentStart, $currentEnd, $nonPersonalScope));
     }
 
     private function buildRecordsStat(
@@ -160,7 +148,7 @@ final class PlatformGrowthStatsWidget extends StatsOverviewWidget
     /**
      * @return array<int, int>
      */
-    private function buildSparkline(string $modelClass, CarbonImmutable $start, CarbonImmutable $end, ?\Closure $scope = null): array
+    private function buildModelSparkline(string $modelClass, CarbonImmutable $start, CarbonImmutable $end, ?\Closure $scope = null): array
     {
         $days = (int) $start->diffInDays($end);
         $points = min($days, 7);
@@ -185,19 +173,7 @@ final class PlatformGrowthStatsWidget extends StatsOverviewWidget
             $scope($query);
         }
 
-        $rows = $query->get();
-
-        $buckets = array_fill(0, $points, 0);
-
-        foreach ($rows as $row) {
-            $idx = min((int) $row->bucket, $points - 1);
-
-            if ($idx >= 0) {
-                $buckets[$idx] += (int) $row->cnt;
-            }
-        }
-
-        return $buckets;
+        return $this->fillBuckets($query->get(), $points);
     }
 
     /**
@@ -230,41 +206,6 @@ final class PlatformGrowthStatsWidget extends StatsOverviewWidget
 
         $rows = DB::select($sql, [$start->toDateTimeString(), $segmentSeconds, ...$bindings]);
 
-        $buckets = array_fill(0, $points, 0);
-
-        foreach ($rows as $row) {
-            $idx = min((int) $row->bucket, $points - 1);
-
-            if ($idx >= 0) {
-                $buckets[$idx] += (int) $row->cnt;
-            }
-        }
-
-        return $buckets;
-    }
-
-    private function bucketExpression(): string
-    {
-        return 'FLOOR(EXTRACT(EPOCH FROM ("created_at" - ?::timestamp)) / ?)';
-    }
-
-    private function calculateChange(int $current, int $previous): float
-    {
-        if ($previous === 0) {
-            return $current > 0 ? 100.0 : 0.0;
-        }
-
-        return round((($current - $previous) / $previous) * 100, 1);
-    }
-
-    private function formatChange(float $change): string
-    {
-        if ($change === 0.0) {
-            return '';
-        }
-
-        $sign = $change > 0 ? '+' : '';
-
-        return " ({$sign}{$change}%)";
+        return $this->fillBuckets(collect($rows), $points);
     }
 }
